Updated Makefile for NT to include generalized arrays stuff.

+++ src/arch/win32/sockRoutines.h
@@ -9,6 +9,6 @@
 
 
 unsigned int skt_ip();
-void skt_server(unsigned int *pip,unsigned int *ppo,unsigned int *pfd);
+void skt_server(unsigned int *pip,unsigned short *ppo,unsigned int *pfd);
 void skt_accept(int src,unsigned int *pip,unsigned int *ppo,unsigned int *pfd);
 int skt_connect(unsigned int ip, int port);
